+++ src/index.js
@@ -1,7 +1,17 @@
+// Import Styles
+
+import './index.css';
+
+// Import React Dependences
+
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+
+// Import Page Component
+
 import App from './App';
+
+// Initialize React Page
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
