+++ src/pages/home/home.jsx
@@ -6,23 +6,27 @@ import "./home.css"
 
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { useState } from "react";
-
-// Components Imports
-
-import Results from '../../components/results/results';
-import SearchBar from '../../components/search/search';
-import LandingPage from "../../components/landing/landing";
-import NoData from "../../components/search/nodata/nodata";
 import ReactLoading from 'react-loading';
 import { Col } from "react-bootstrap";
 
-// Page Home HTML Code
+// Page Components Imports
+
+import Results from './results/results';
+import SearchBar from '../../components/search/search';
+import LandingPage from './landing/landing'
+import NoData from "./nodata/nodata";
+
+// Page Home
 
 function Home(){
+
+    // Initialize useState variables
 
     const [Searched, setSearched] = useState('');
     const [Dados, setDados] = useState([]);
     const [LanguageProg, setLanguageProg] = useState('javascript');
+
+    // HTML Code
 
     return (
         <div className="App">
