+++ src/pages/home/results/card/card.css
