+++ src/pages/perfil/createdFunctions/card/card.css
