+++ src/pages/perfil/savedFunctions/savedFunctions.jsx
@@ -1,55 +1,74 @@
-import "./createdFunctions.css"
+// Import Styles
 
-import { supabase } from "../../providers/supabase";
-import { useEffect, useState } from "react";
-import Card from "./card/card";
+import "./savedFunctions.css"
 
+// Import Frameworks
 
-function CreatedFunctions(){
+import { supabase } from "../../../providers/supabase"
+import { useEffect, useState } from "react"
 
+// Screen Components
+
+import Card from "./card/card"
+
+// Component Funciton
+
+function SavedFunctions(){
+
+    // Initialize useState Variables
+
+    const [ functions, setFunctions ] = useState([])
     const [ mainLanguage, setMainLanguage ] = useState('javascript');
     const [ codeLanguage, setCodeLanguage ] = useState('1');
-    const [ functions, setFunctions ] = useState([])
     const [ dataSecao, setDataSecao ] = useState([])
+
+    // Take Session user data
 
     useEffect(() => {
         async function obterUsuarioSessao(){
             const { data, error } = await supabase.auth.getSession()
             setDataSecao(data)
 
-            const { data: data2, error: error2 } = await supabase.from('1functions').select('*').in('id', data.session.user.user_metadata.criadas)
+            const { data: data2, error: error2 } = await supabase.from('1functions').select('*').in('id', data.session.user.user_metadata.salvas)
+
             setFunctions(data2)
         }
         obterUsuarioSessao()
     },[])
 
+    // Take saved functions from session user
+
     useEffect(() => {
         if (!dataSecao.session){return}
         async function obterFuncoes(){
-            const { data, error } = await supabase.from(codeLanguage+'functions').select('*').in('id', dataSecao.session.user.user_metadata.criadas)
+            const { data, error } = await supabase.from(codeLanguage+'functions').select('*').in('id', dataSecao.session.user.user_metadata.salvas)
             setFunctions(data)
         }
         obterFuncoes()
     },[mainLanguage])
+
+    // Change Language Programming
 
     async function changeLanguage(){
 
         var language = document.getElementsByName('selector')
         for (var i = 0; i < language.length; i++) {
             if (language[i].checked) {
-              var linguagem = language[i].value
-              let { data: languages, error } = await supabase.from('planguages').select('id').eq('planguage',linguagem)
+                var linguagem = language[i].value
+                let { data: languages, error } = await supabase.from('planguages').select('id').eq('planguage',linguagem)
 
-              setMainLanguage(linguagem)
-              setCodeLanguage(languages[0].id)
-              break
+                setMainLanguage(linguagem)
+                setCodeLanguage(languages[0].id)
+                break
             }
         }
     }
 
-    return (
-        <div className="createdFunctionsMain">
+    // HTML Code
 
+    return (
+        <div className="savedFunctionsMain">
+            
             <div class="radio-selector" id="radioSalvas">
                 <input type="radio" id="option1" name="selector" value="javascript" onChange={changeLanguage} />
                 <label for="option1">javascript</label>
@@ -61,9 +80,10 @@ function CreatedFunctions(){
                 <Card key={element.id} functionData={element} language={mainLanguage} />
             ))}
 
+
         </div>
     )
 
 }
 
-export default CreatedFunctions;
+export default SavedFunctions;
