+++ src/pages/sobre/card/card.css
