+++ src/pages/sobre/sobre.css
@@ -1,0 +1,27 @@
+#grid{
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+}
+#titleTeam{
+    color: white;
+}
+#boxTitle{
+    background-color: black;
+    height: 60px;
+    width: 250px;
+    border-radius: 10px;
+    box-shadow: 0px 0px 10px black;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+#titlePosition{
+    display: flex;
+    justify-content: center;
+    width: 100%;
+    align-items: center;
+    margin-top: 150px;
+}
